refactor : 각 클래스들을 패키지로 분류
